Init Bokeh_Util.js + Functional WellCapture App
